Implemented rectangle selection tool
This is a rudimentary implementation of a rectangle selection tool.
Now it is possible to select all objects in a given rectangle at
once and move or delete them together. The selection is done via
the bounding boxes of the elements, which is kind of clumsy. Also
there is no indicator of the selection rectangle Some improvement
on this can surely be made. Since we now can selectively remove
parts of the drawing, this closes #2.