Introducing the panning tool
This introduces a pan tool, which is used to pan. Like two
fingered scrolling and the middle mouse click. The CTRL to pan
feature however has been removed, since it is not behaving in
the same way over different operating systems. Modifier keys
are behaving different in every browser and OS.

It is a nightmare to unify input controls over the different
platforms. Things that work in Windows do not work on macOS or a
GNU/Linux distribution. Since we don't want to break workflow when
switching the OS, we minimize keyboard and mouse inputs to the
stuff that does work on all platforms. Furthermore the interface
should be usable by workstation featuring a mouse and ONE button.
Meaning all features should be accessible through selectable GUI
buttons.